Exception handling

+++ src/main/java/com/devsuperior/dscommerce/Services/ProductService.java
@@ -1,15 +1,20 @@
 package com.devsuperior.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dtos.ProductDto;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
 import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -41,10 +46,14 @@ public class ProductService {
 
     @Transactional
     public ProductDto update(Long id, ProductDto productDto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(productDto, entity);
-        entity = repository.save(entity);
-        return new ProductDto(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(productDto, entity);
+            entity = repository.save(entity);
+            return new ProductDto(entity);
+        } catch (EntityNotFoundException e ) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     private void copyDtoToEntity(ProductDto dto, Product entity) {
@@ -54,10 +63,18 @@ public class ProductService {
         entity.setImgUrl(dto.getImgUrl());
     }
 
-    
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
-    }
+	if (!repository.existsById(id)) {
+		throw new ResourceNotFoundException("Recurso não encontrado");
+	}
+	try {
+        repository.deleteById(id);    		
+	}
+    catch (DataIntegrityViolationException e) {
+        throw new DatabaseException("Falha de integridade referencial");
+   	}
+}
+
     
 }
